[NODERS]TEAM gentx
[NODERS]TEAM is part of the developing ecosystem Goshu Alliance aimed at developing the community in various crypto projects (From projects with nodes to various game projects)

[NODERS]TEAM - a team of professional developers with more than five years of experience in the IT field.
At the moment, our main goal is to participate in various testnets of projects built on different networks and help in building a strong community.

We help both with the organization of testnets and with their support (Moderation, assistance in writing technical documentation, helping the community get acquainted with the project, writing guides, holding AMAs, etc.)

Also, one more of our global goals is to automate the validator management process and simplify it by writing an application for this process, as well as further growth into a large staking platform.

This application will perform the function of writing commands on the server, but through telephones using predefined commands for a specific project. You will only need to click on the button you need (It will increase the stake, withdraw rewards, vote, check the status of the node and the validator, etc.)

We are also engaged in writing statesyncs, relayers, snapshots and everything that can simplify the installation of a node and track its state.

We are currently participating in several public and closed testnets. We have several nodes in the mainnet with stakes from projects, and we are also writing a large website where all our developments for testnets will be collected, all projects in which we participate and will participate, guides, statesinks, bots for TG and much more

About us:
💻 Site https://noders-stake.com/
💻 My GitHub https://github.com/nodersteam
📝 TG channel https://t.me/nodersteam